// src/app/shared/model/more-address.model.ts
export class MoreAddress
{
    countries: string[]=[];
    states: string[]=[];
    cities: string[]=[];

    selectedCountryId:number=0;
    selectedCountry: string='';
    selectedStateId:number=0;
    selectedState: string='';
    selectedCity: string='';

    pincodeText:string = '';
}